Refactored Project structure
- CSVReaderUtility created to load the CSV file and map to UserBean.
  Test data kept in src/test/resources
