+++ src/test/java/com/api/utils/CSVReaderUtility.java
@@ -1,25 +1,29 @@
-package com.demo.csv;
+package com.api.utils;
 
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.List;
 
+import com.dataproviders.api.bean.UserBean;
 import com.opencsv.CSVReader;
 import com.opencsv.bean.CsvToBean;
 import com.opencsv.bean.CsvToBeanBuilder;
-import com.opencsv.exceptions.CsvException;
 
-public class ReadCSVFileMapToPojo {
+public class CSVReaderUtility {
 
-	public static void main(String[] args) throws IOException, CsvException {
+	private CSVReaderUtility() {
+		
+	}
+	public static void loadCSV(String pathOfCSVFile) {
 
 		InputStream inputStream = Thread.currentThread().getContextClassLoader()
-				.getResourceAsStream("testData/LoginCreds.csv");
+				.getResourceAsStream(pathOfCSVFile);
 		InputStreamReader inputStreamReader = new InputStreamReader(inputStream);
 		CSVReader csvReader = new CSVReader(inputStreamReader);
 
-		CsvToBean<UserBean> csvBean = new CsvToBeanBuilder(csvReader).withType(UserBean.class).withIgnoreEmptyLine(true)
+		CsvToBean<UserBean> csvBean = new CsvToBeanBuilder(csvReader)
+				.withType(UserBean.class)
+				.withIgnoreEmptyLine(true)
 				.build();
 		List<UserBean> userlist = csvBean.parse();
 		System.out.println(userlist);
