+++ src/test/java/com/dataproviders/api/bean/UserBean.java
@@ -1,4 +1,4 @@
-package com.demo.csv;
+package com.dataproviders.api.bean;
 
 public class UserBean {
 	private String username;
